Jitter Sound Player :)
Check out the new node! it's in action on the player scene. you can play a sound effect with some built in pitch jitter. listen to the laser sfx and see that it slightly changes pitch each time it's played so it doesn't get too monotonous.

Also, the speech label supports it as well now.