Generic Bullets
Made Bullet Generic. Updates to Spawner component to accomodate, now takes a Callable as an optional parameter when spawning. The Callable is called with the spawned node as its parameter, so we can do things to the bullet before it gets added to the scene tree (such as setting its damage, direction, etc)